+++ src/Resources/contao/config/config.php
@@ -33,11 +33,20 @@ $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE']['hh_newsalert'] = array
     ]
 );
 
-$GLOBALS['BE_MOD']['content']['newsalert'] = [
-    'tables' => ['tl_newsalert_recipients']
-];
+//$GLOBALS['BE_MOD']['content']['newsalert'] = [
+//    'tables' => ['tl_newsalert_recipients']
+//];
+
+$GLOBALS['BE_MOD']['content']['news']['tables'][] = 'tl_newsalert_recipients';
 
 
-
-//array_insert($GLOBALS['BE_MOD']['content']['news']['tables'])
+//array_insert($GLOBALS['BE_MOD']['content']['news']['tables'], 1, array
+//(
+//    array('tl_newsalert_recipients'),
+//));
+//
+//array_insert();
+//
+//$GLOBALS['BE_MOD']['content']['news']['tables'] =
+//    array_merge($GLOBALS['BE_MOD']['content']['news']['tables'], ['tl_newsalert_recipients']);