+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 * Recipients table (dca)
-* Backend module for recipients
+* Backend module for recipients (placed as global operation under news)
 
 ### Fixed
 * content element illegal offset